--- a/framework/app.mk
+++ b/framework/app.mk
@@ -80,8 +80,6 @@ $(eval $(call unity_dir_rule, $(unity_src_dir)))
 # that won't benefit from unity building
 # Also, exclude the base directory by default because it's another big jumble
 # of unrelated stuff.
-# Note: If a top-level directory is excluded from being built with unity, we
-# ignore contents of all subdirectories under that folder.
 non_unity_dirs := %.libs %/src $(app_non_unity_dirs)
 
 # Find all of the top-level subdirectories in our src folder(s)
@@ -89,10 +87,11 @@ non_unity_dirs := %.libs %/src $(app_non_unity_dirs)
 # The idea is that files grouped withing a subdirectory are closely related
 # and will benefit from a Unity build
 srcsubdirs := $(shell find $(APPLICATION_DIR)/src -maxdepth 1 -type d -not -path '*/.libs*')
+allsrcsubdirs := $(shell find $(APPLICATION_DIR)/src -type d -not -path '*/.libs*')
 
 # Filter out the paths we don't want to Unity build
 unity_srcsubdirs := $(filter-out $(non_unity_dirs), $(srcsubdirs))
-non_unity_srcsubdirs := $(filter $(non_unity_dirs), $(srcsubdirs))
+non_unity_srcsubdirs := $(filter $(non_unity_dirs), $(allsrcsubdirs))
 
 # This is a biggie
 # Loop over the subdirectories, creating a rule to create the Unity source file