By default lump all source at each top level directory into single unity files. This is a departure from before when we did one unity file per directory.

closes #26028