@dschwen 2023-11-14
This is not going to work. We make use of subdirectories deliberately to avoid lumping stuff in with the rest of the unity sources for libtorch for example.