@permcody 2023-11-14
The capability to remove individual directories from being “unitied” still exists. Can we not just add those directories explicitly to the list where needed?