@permcody 2023-12-20
Timings

Before: `make -j10 1286.05s user 196.03s system 569% cpu 4:20.24 total`
After: `make -j10 691.59s user 117.54s system 464% cpu 2:54:05 total`