@permcody 2024-01-08
I've opened issue https://github.com/idaholab/moose/issues/26508 to look at the revision generation issue separately. It is not related to this PR.