@permcody 2024-01-09
> This is not going to work.

Narrator: ... but it did work, and they all lived happily ever after!